+++ cmd/gaia/app/export.go
@@ -13,7 +13,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/gov"
 	"github.com/cosmos/cosmos-sdk/x/mint"
 	"github.com/cosmos/cosmos-sdk/x/slashing"
-	stake "github.com/cosmos/cosmos-sdk/x/stake"
+	"github.com/cosmos/cosmos-sdk/x/staking"
 )
 
 // export the state of gaia for a genesis file
@@ -39,7 +39,7 @@ func (app *GaiaApp) ExportAppStateAndValidators(forZeroHeight bool) (
 	genState := NewGenesisState(
 		accounts,
 		auth.ExportGenesis(ctx, app.accountKeeper, app.feeCollectionKeeper),
-		stake.ExportGenesis(ctx, app.stakeKeeper),
+		staking.ExportGenesis(ctx, app.stakingKeeper),
 		mint.ExportGenesis(ctx, app.mintKeeper),
 		distr.ExportGenesis(ctx, app.distrKeeper),
 		gov.ExportGenesis(ctx, app.govKeeper),
@@ -49,7 +49,7 @@ func (app *GaiaApp) ExportAppStateAndValidators(forZeroHeight bool) (
 	if err != nil {
 		return nil, nil, err
 	}
-	validators = stake.WriteValidators(ctx, app.stakeKeeper)
+	validators = staking.WriteValidators(ctx, app.stakingKeeper)
 	return appState, validators, nil
 }
 
@@ -62,13 +62,13 @@ func (app *GaiaApp) prepForZeroHeightGenesis(ctx sdk.Context) {
 	/* Handle fee distribution state. */
 
 	// withdraw all validator commission
-	app.stakeKeeper.IterateValidators(ctx, func(_ int64, val sdk.Validator) (stop bool) {
+	app.stakingKeeper.IterateValidators(ctx, func(_ int64, val sdk.Validator) (stop bool) {
 		_ = app.distrKeeper.WithdrawValidatorCommission(ctx, val.GetOperator())
 		return false
 	})
 
 	// withdraw all delegator rewards
-	dels := app.stakeKeeper.GetAllDelegations(ctx)
+	dels := app.stakingKeeper.GetAllDelegations(ctx)
 	for _, delegation := range dels {
 		_ = app.distrKeeper.WithdrawDelegationRewards(ctx, delegation.DelegatorAddr, delegation.ValidatorAddr)
 	}
@@ -84,7 +84,7 @@ func (app *GaiaApp) prepForZeroHeightGenesis(ctx sdk.Context) {
 	ctx = ctx.WithBlockHeight(0)
 
 	// reinitialize all validators
-	app.stakeKeeper.IterateValidators(ctx, func(_ int64, val sdk.Validator) (stop bool) {
+	app.stakingKeeper.IterateValidators(ctx, func(_ int64, val sdk.Validator) (stop bool) {
 		app.distrKeeper.Hooks().AfterValidatorCreated(ctx, val.GetOperator())
 		return false
 	})
@@ -97,32 +97,32 @@ func (app *GaiaApp) prepForZeroHeightGenesis(ctx sdk.Context) {
 	// reset context height
 	ctx = ctx.WithBlockHeight(height)
 
-	/* Handle stake state. */
+	/* Handle staking state. */
 
 	// iterate through redelegations, reset creation height
-	app.stakeKeeper.IterateRedelegations(ctx, func(_ int64, red stake.Redelegation) (stop bool) {
+	app.stakingKeeper.IterateRedelegations(ctx, func(_ int64, red staking.Redelegation) (stop bool) {
 		red.CreationHeight = 0
-		app.stakeKeeper.SetRedelegation(ctx, red)
+		app.stakingKeeper.SetRedelegation(ctx, red)
 		return false
 	})
 
 	// iterate through unbonding delegations, reset creation height
-	app.stakeKeeper.IterateUnbondingDelegations(ctx, func(_ int64, ubd stake.UnbondingDelegation) (stop bool) {
+	app.stakingKeeper.IterateUnbondingDelegations(ctx, func(_ int64, ubd staking.UnbondingDelegation) (stop bool) {
 		ubd.CreationHeight = 0
-		app.stakeKeeper.SetUnbondingDelegation(ctx, ubd)
+		app.stakingKeeper.SetUnbondingDelegation(ctx, ubd)
 		return false
 	})
 
 	// Iterate through validators by power descending, reset bond heights, and
 	// update bond intra-tx counters.
-	store := ctx.KVStore(app.keyStake)
-	iter := sdk.KVStoreReversePrefixIterator(store, stake.ValidatorsKey)
+	store := ctx.KVStore(app.keyStaking)
+	iter := sdk.KVStoreReversePrefixIterator(store, staking.ValidatorsKey)
 	counter := int16(0)
 
 	var valConsAddrs []sdk.ConsAddress
 	for ; iter.Valid(); iter.Next() {
 		addr := sdk.ValAddress(iter.Key()[1:])
-		validator, found := app.stakeKeeper.GetValidator(ctx, addr)
+		validator, found := app.stakingKeeper.GetValidator(ctx, addr)
 		if !found {
 			panic("expected validator, not found")
 		}
@@ -131,26 +131,13 @@ func (app *GaiaApp) prepForZeroHeightGenesis(ctx sdk.Context) {
 		validator.UnbondingHeight = 0
 		valConsAddrs = append(valConsAddrs, validator.ConsAddress())
 
-		app.stakeKeeper.SetValidator(ctx, validator)
+		app.stakingKeeper.SetValidator(ctx, validator)
 		counter++
 	}
 
 	iter.Close()
 
 	/* Handle slashing state. */
-
-	// remove all existing slashing periods and recreate one for each validator
-	app.slashingKeeper.DeleteValidatorSlashingPeriods(ctx)
-
-	for _, valConsAddr := range valConsAddrs {
-		sp := slashing.ValidatorSlashingPeriod{
-			ValidatorAddr: valConsAddr,
-			StartHeight:   0,
-			EndHeight:     0,
-			SlashedSoFar:  sdk.ZeroDec(),
-		}
-		app.slashingKeeper.SetValidatorSlashingPeriod(ctx, sp)
-	}
 
 	// reset start height on signing infos
 	app.slashingKeeper.IterateValidatorSigningInfos(