+++ docs/examples/democoin/x/simplestaking/handler.go
@@ -1,0 +1,33 @@
+package simplestaking
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// NewHandler returns a handler for "simplestaking" type messages.
+func NewHandler(k Keeper) sdk.Handler {
+	return func(ctx sdk.Context, msg sdk.Msg) sdk.Result {
+		switch msg.(type) {
+		case MsgBond:
+			return handleMsgBond()
+		case MsgUnbond:
+			return handleMsgUnbond()
+		default:
+			return sdk.ErrUnknownRequest("No match for message type.").Result()
+		}
+	}
+}
+
+func handleMsgBond() sdk.Result {
+	// Removed ValidatorSet from result because it does not get used.
+	// TODO: Implement correct bond/unbond handling
+	return sdk.Result{
+		Code: sdk.CodeOK,
+	}
+}
+
+func handleMsgUnbond() sdk.Result {
+	return sdk.Result{
+		Code: sdk.CodeOK,
+	}
+}