+++ server/tm_cmds.go
@@ -41,8 +41,10 @@ func ShowValidatorCmd(ctx *Context) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 
 			cfg := ctx.Config
-			privValidator := pvm.LoadOrGenFilePV(cfg.PrivValidatorFile())
-			valPubKey := privValidator.PubKey
+			UpgradeOldPrivValFile(cfg)
+			privValidator := pvm.LoadOrGenFilePV(
+				cfg.PrivValidatorKeyFile(), cfg.PrivValidatorStateFile())
+			valPubKey := privValidator.GetPubKey()
 
 			if viper.GetBool(client.FlagJson) {
 				return printlnJSON(valPubKey)
@@ -67,9 +69,12 @@ func ShowAddressCmd(ctx *Context) *cobra.Command {
 		Use:   "show-address",
 		Short: "Shows this node's tendermint validator consensus address",
 		RunE: func(cmd *cobra.Command, args []string) error {
+
 			cfg := ctx.Config
-			privValidator := pvm.LoadOrGenFilePV(cfg.PrivValidatorFile())
-			valConsAddr := (sdk.ConsAddress)(privValidator.Address)
+			UpgradeOldPrivValFile(cfg)
+			privValidator := pvm.LoadOrGenFilePV(
+				cfg.PrivValidatorKeyFile(), cfg.PrivValidatorStateFile())
+			valConsAddr := (sdk.ConsAddress)(privValidator.GetAddress())
 
 			if viper.GetBool(client.FlagJson) {
 				return printlnJSON(valConsAddr)
@@ -102,7 +107,7 @@ func UnsafeResetAllCmd(ctx *Context) *cobra.Command {
 		Short: "Resets the blockchain database, removes address book files, and resets priv_validator.json to the genesis state",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cfg := ctx.Config
-			tcmd.ResetAll(cfg.DBDir(), cfg.P2P.AddrBookFile(), cfg.PrivValidatorFile(), ctx.Logger)
+			tcmd.ResetAll(cfg.DBDir(), cfg.P2P.AddrBookFile(), cfg.PrivValidatorKeyFile(), cfg.PrivValidatorStateFile(), ctx.Logger)
 			return nil
 		},
 	}