+++ x/distribution/keeper/calculation_test.go
@@ -6,24 +6,24 @@ import (
 	"github.com/stretchr/testify/require"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/stake"
+	"github.com/cosmos/cosmos-sdk/x/staking"
 )
 
 func TestCalculateRewardsBasic(t *testing.T) {
 	ctx, _, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// fetch validator and delegation
 	val := sk.Validator(ctx, valOpAddr1)
@@ -39,7 +39,7 @@ func TestCalculateRewardsBasic(t *testing.T) {
 	require.True(t, rewards.IsZero())
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// end period
@@ -49,27 +49,27 @@ func TestCalculateRewardsBasic(t *testing.T) {
 	rewards = k.calculateDelegationRewards(ctx, val, del, endingPeriod)
 
 	// rewards should be half the tokens
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(5)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(5)}}, rewards)
 
 	// commission should be the other half
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(5)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(5)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
 }
 
 func TestCalculateRewardsAfterSlash(t *testing.T) {
 	ctx, _, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// fetch validator and delegation
 	val := sk.Validator(ctx, valOpAddr1)
@@ -97,7 +97,7 @@ func TestCalculateRewardsAfterSlash(t *testing.T) {
 	ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 3)
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// end period
@@ -107,27 +107,27 @@ func TestCalculateRewardsAfterSlash(t *testing.T) {
 	rewards = k.calculateDelegationRewards(ctx, val, del, endingPeriod)
 
 	// rewards should be half the tokens
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(5)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(5)}}, rewards)
 
 	// commission should be the other half
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(5)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(5)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
 }
 
 func TestCalculateRewardsAfterManySlashes(t *testing.T) {
 	ctx, _, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// fetch validator and delegation
 	val := sk.Validator(ctx, valOpAddr1)
@@ -155,7 +155,7 @@ func TestCalculateRewardsAfterManySlashes(t *testing.T) {
 	ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 3)
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// slash the validator by 50% again
@@ -177,38 +177,38 @@ func TestCalculateRewardsAfterManySlashes(t *testing.T) {
 	rewards = k.calculateDelegationRewards(ctx, val, del, endingPeriod)
 
 	// rewards should be half the tokens
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}, rewards)
 
 	// commission should be the other half
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
 }
 
 func TestCalculateRewardsMultiDelegator(t *testing.T) {
 	ctx, _, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// fetch validator and delegation
 	val := sk.Validator(ctx, valOpAddr1)
 	del1 := sk.Delegation(ctx, sdk.AccAddress(valOpAddr1), valOpAddr1)
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// second delegation
-	msg2 := stake.NewMsgDelegate(sdk.AccAddress(valOpAddr2), valOpAddr1, sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)))
+	msg2 := staking.NewMsgDelegate(sdk.AccAddress(valOpAddr2), valOpAddr1, sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)))
 	require.True(t, sh(ctx, msg2).IsOK())
 	del2 := sk.Delegation(ctx, sdk.AccAddress(valOpAddr2), valOpAddr1)
 
@@ -216,7 +216,7 @@ func TestCalculateRewardsMultiDelegator(t *testing.T) {
 	val = sk.Validator(ctx, valOpAddr1)
 
 	// end block
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// allocate some more rewards
 	k.AllocateTokensToValidator(ctx, val, tokens)
@@ -228,36 +228,36 @@ func TestCalculateRewardsMultiDelegator(t *testing.T) {
 	rewards := k.calculateDelegationRewards(ctx, val, del1, endingPeriod)
 
 	// rewards for del1 should be 7.5
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDecWithPrec(75, 1)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDecWithPrec(75, 1)}}, rewards)
 
 	// calculate delegation rewards for del2
 	rewards = k.calculateDelegationRewards(ctx, val, del2, endingPeriod)
 
 	// rewards for del2 should be 2.5
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDecWithPrec(25, 1)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDecWithPrec(25, 1)}}, rewards)
 
 	// commission should be 10
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
 }
 
 func TestWithdrawDelegationRewardsBasic(t *testing.T) {
 	ctx, ak, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// assert correct initial balance
-	require.Equal(t, sdk.Coins{{stake.DefaultBondDenom, sdk.NewInt(900)}}, ak.GetAccount(ctx, sdk.AccAddress(valOpAddr1)).GetCoins())
+	require.Equal(t, sdk.Coins{{staking.DefaultBondDenom, sdk.NewInt(900)}}, ak.GetAccount(ctx, sdk.AccAddress(valOpAddr1)).GetCoins())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// set zero outstanding rewards
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
@@ -266,37 +266,37 @@ func TestWithdrawDelegationRewardsBasic(t *testing.T) {
 	val := sk.Validator(ctx, valOpAddr1)
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// withdraw rewards
 	require.Nil(t, k.WithdrawDelegationRewards(ctx, sdk.AccAddress(valOpAddr1), valOpAddr1))
 
 	// assert correct balance
-	require.Equal(t, sdk.Coins{{stake.DefaultBondDenom, sdk.NewInt(905)}}, ak.GetAccount(ctx, sdk.AccAddress(valOpAddr1)).GetCoins())
+	require.Equal(t, sdk.Coins{{staking.DefaultBondDenom, sdk.NewInt(905)}}, ak.GetAccount(ctx, sdk.AccAddress(valOpAddr1)).GetCoins())
 
 	// withdraw commission
 	require.Nil(t, k.WithdrawValidatorCommission(ctx, valOpAddr1))
 
 	// assert correct balance
-	require.Equal(t, sdk.Coins{{stake.DefaultBondDenom, sdk.NewInt(910)}}, ak.GetAccount(ctx, sdk.AccAddress(valOpAddr1)).GetCoins())
+	require.Equal(t, sdk.Coins{{staking.DefaultBondDenom, sdk.NewInt(910)}}, ak.GetAccount(ctx, sdk.AccAddress(valOpAddr1)).GetCoins())
 }
 
 func TestCalculateRewardsAfterManySlashesInSameBlock(t *testing.T) {
 	ctx, _, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// fetch validator and delegation
 	val := sk.Validator(ctx, valOpAddr1)
@@ -315,7 +315,7 @@ func TestCalculateRewardsAfterManySlashesInSameBlock(t *testing.T) {
 	ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 3)
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// slash the validator by 50%
@@ -340,34 +340,34 @@ func TestCalculateRewardsAfterManySlashesInSameBlock(t *testing.T) {
 	rewards = k.calculateDelegationRewards(ctx, val, del, endingPeriod)
 
 	// rewards should be half the tokens
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}, rewards)
 
 	// commission should be the other half
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(10)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(10)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
 }
 
 func TestCalculateRewardsMultiDelegatorMultiSlash(t *testing.T) {
 	ctx, _, k, sk, _ := CreateTestInputDefault(t, false, 1000)
-	sh := stake.NewHandler(sk)
+	sh := staking.NewHandler(sk)
 
 	// initialize state
 	k.SetOutstandingRewards(ctx, sdk.DecCoins{})
 
 	// create validator with 50% commission
-	commission := stake.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
-	msg := stake.NewMsgCreateValidator(valOpAddr1, valConsPk1,
-		sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)), stake.Description{}, commission)
+	commission := staking.NewCommissionMsg(sdk.NewDecWithPrec(5, 1), sdk.NewDecWithPrec(5, 1), sdk.NewDec(0))
+	msg := staking.NewMsgCreateValidator(valOpAddr1, valConsPk1,
+		sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)), staking.Description{}, commission)
 	require.True(t, sh(ctx, msg).IsOK())
 
 	// end block to bond validator
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// fetch validator and delegation
 	val := sk.Validator(ctx, valOpAddr1)
 	del1 := sk.Delegation(ctx, sdk.AccAddress(valOpAddr1), valOpAddr1)
 
 	// allocate some rewards
-	tokens := sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(30)}}
+	tokens := sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(30)}}
 	k.AllocateTokensToValidator(ctx, val, tokens)
 
 	// slash the validator
@@ -376,12 +376,12 @@ func TestCalculateRewardsMultiDelegatorMultiSlash(t *testing.T) {
 	ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 3)
 
 	// second delegation
-	msg2 := stake.NewMsgDelegate(sdk.AccAddress(valOpAddr2), valOpAddr1, sdk.NewCoin(stake.DefaultBondDenom, sdk.NewInt(100)))
+	msg2 := staking.NewMsgDelegate(sdk.AccAddress(valOpAddr2), valOpAddr1, sdk.NewCoin(staking.DefaultBondDenom, sdk.NewInt(100)))
 	require.True(t, sh(ctx, msg2).IsOK())
 	del2 := sk.Delegation(ctx, sdk.AccAddress(valOpAddr2), valOpAddr1)
 
 	// end block
-	stake.EndBlocker(ctx, sk)
+	staking.EndBlocker(ctx, sk)
 
 	// allocate some more rewards
 	k.AllocateTokensToValidator(ctx, val, tokens)
@@ -401,14 +401,14 @@ func TestCalculateRewardsMultiDelegatorMultiSlash(t *testing.T) {
 	rewards := k.calculateDelegationRewards(ctx, val, del1, endingPeriod)
 
 	// rewards for del1 should be 20 (15 + 5)
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDecWithPrec(20, 0)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDecWithPrec(20, 0)}}, rewards)
 
 	// calculate delegation rewards for del2
 	rewards = k.calculateDelegationRewards(ctx, val, del2, endingPeriod)
 
 	// rewards for del2 should be 10
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDecWithPrec(10, 0)}}, rewards)
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDecWithPrec(10, 0)}}, rewards)
 
 	// commission should be 30 (unaffected by slashing)
-	require.Equal(t, sdk.DecCoins{{stake.DefaultBondDenom, sdk.NewDec(30)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
+	require.Equal(t, sdk.DecCoins{{staking.DefaultBondDenom, sdk.NewDec(30)}}, k.GetValidatorAccumulatedCommission(ctx, valOpAddr1))
 }