+++ CHANGELOG.md
@@ -134,7 +134,7 @@ BREAKING CHANGES
   * [\#2019](https://github.com/cosmos/cosmos-sdk/issues/2019) Cap total number of signatures. Current per-transaction limit is 7, and if that is exceeded transaction is rejected.
   * [\#2801](https://github.com/cosmos/cosmos-sdk/pull/2801) Remove AppInit structure.
   * [\#2798](https://github.com/cosmos/cosmos-sdk/issues/2798) Governance API has miss-spelled English word in JSON response ('depositer' -> 'depositor')
-  * [\#2943](https://github.com/cosmos/cosmos-sdk/pull/2943) Transaction action tags equal the message type. Stake EndBlocker tags are included.
+  * [\#2943](https://github.com/cosmos/cosmos-sdk/pull/2943) Transaction action tags equal the message type. Staking EndBlocker tags are included.
 
 * Tendermint
   * Update to Tendermint 0.27.0
@@ -1062,7 +1062,7 @@ BREAKING CHANGES
 * All module keepers now require a codespace, see basecoin or democoin for usage
 * Many changes to names throughout
   * Type as a prefix naming convention applied (ex. BondMsg -> MsgBond)
-  * Removed redundancy in names (ex. stake.StakeKeeper -> stake.Keeper)
+  * Removed redundancy in names (ex. stake.StakingKeeper -> stake.Keeper)
 * Removed SealedAccountMapper
 * gaiad init now requires use of `--name` flag
 * Removed Get from Msg interface